+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 144-hf1:
 
 * Update reports for hybrid assessment
+* Handle invalid deflate frames
 
 --------------------------------------------------------------------------------